'Removed whitespace at the end of base_converter_cog and moved eight_ball_cog to memes_cog'
