+++ bot/cogs/base_converter_cog.py
@@ -127,7 +127,3 @@ class BaseConverterCog(commands.Cog):
 
 def setup(bot):
     bot.add_cog(BaseConverterCog(bot))
-
-
-
-
